+++ .gitignore
@@ -27,3 +27,5 @@ hs_err_pid*
 replay_pid*
 
 .mvn/
+
+src/test/resources/*.json
