+++ README.md
@@ -1,2 +1,6 @@
 # prayingtimes-calculator
 Java port of Prayer Times Calculator from PrayTimes.org
+
+Tested to work only for Spiritual Administration of Muslims of Russia Method
+
+Before build run 'node prepareTestData.js' in 'src/test/resources/' to prepare test data
