--- src/main/java/ru/mfilatov/constants/SunPositionOffsets.java
@@ -1,9 +1,0 @@
-package ru.mfilatov.constants;
-
-public class SunPositionOffsets {
-  public static final double FAJR = 5.0;
-  public static final double SUNRISE = 6.0;
-  public static final double DHUHR = 12.0;
-  public static final double ASR = 13.0;
-  public static final double SUNSET = 18.0;
-}
