--- src/main/java/ru/mfilatov/enums/AdjustForHigherLatitudes.java
@@ -1,8 +1,0 @@
-package ru.mfilatov.enums;
-
-public enum AdjustForHigherLatitudes {
-  NIGHT_MIDDLE, // middle of night
-  ANGLE_BASED, // angle/60th of night
-  ONE_SEVENTH, // 1/7th of night
-  NONE; // No adjustment
-}
