--- src/main/java/ru/mfilatov/enums/AsrJuristicMethod.java
@@ -1,7 +1,0 @@
-package ru.mfilatov.enums;
-
-/** STANDARD48 // Shafi`i, Maliki, Ja`fari, Hanbali HANAFI // Hanafi */
-public enum AsrJuristicMethod {
-  STANDARD,
-  HANAFI;
-}
