--- src/main/java/ru/mfilatov/enums/MidnightMode.java
@@ -1,6 +1,0 @@
-package ru.mfilatov.enums;
-
-public enum MidnightMode {
-  STANDARD, // Mid Sunset to Sunrise
-  JAFARI; // Mid Sunset to Fajr
-}
