--- src/main/java/ru/mfilatov/model/TimesDouble.java
@@ -1,12 +1,0 @@
-package ru.mfilatov.model;
-
-public record TimesDouble(
-    double imsak,
-    double fajr,
-    double sunrise,
-    double dhuhr,
-    double asr,
-    double sunset,
-    double maghrib,
-    double isha,
-    double midnight) {}
