+++ src/main/java/ru/mfilatov/prayingtimes/calculator/constants/SunPositionOffsets.java
@@ -17,19 +17,12 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-const fs = require('node:fs');
+package ru.mfilatov.prayingtimes.calculator.constants;
 
-eval(fs.readFileSync('PrayTimes.js').toString())
-prayTimes.setMethod("RUSSIA")
-console.log(prayTimes.getTimes(new Date(), [55.75222, 37.61556], 3))
-
-const content = prayTimes.getTimes(new Date(), [55.75222, 37.61556], 3);
-
-
-fs.writeFile('moscow.json', JSON.stringify(content), err => {
-    if (err) {
-      console.error(err);
-    } else {
-      // file written successfully
-    }
-  });
+public class SunPositionOffsets {
+  public static final double FAJR = 5.0;
+  public static final double SUNRISE = 6.0;
+  public static final double DHUHR = 12.0;
+  public static final double ASR = 13.0;
+  public static final double SUNSET = 18.0;
+}
