+++ src/test/java/ru/mfilatov/prayingtimes/calculator/JulianDayCalculatorTest.java
@@ -1,10 +1,30 @@
-package ru.mfilatov;
+/*
+ * PrayingTimes Calculator
+ * Copyright (C) 2024 Mikhail Filatov
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+package ru.mfilatov.prayingtimes.calculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
 import org.junit.jupiter.api.Test;
-import ru.mfilatov.functions.JulianDayCalculator;
+
+import ru.mfilatov.prayingtimes.calculator.functions.JulianDayCalculator;
 
 
 public class JulianDayCalculatorTest {
@@ -23,7 +43,7 @@ public class JulianDayCalculatorTest {
   }
 
   @Test
-  void compareAlgorithmResults() {
+  void compareAlgorithmResultsTest() {
     var calc = new JulianDayCalculator();
     var startDate = LocalDate.now();
     var daysToTest = 365;
